NOW GIF CONTAINER COMPONENT AND GIF COMPONENT MARK THE FAVORITE GIFS
