+++ src/components/GifGrid/GifGrid.jsx
@@ -2,9 +2,15 @@ import React from 'react';
 
 import Gif from '../Gif';
 
-const GifGrid = ({ gifs, filterText }) => (
+const GifGrid = ({ gifs, favoriteGifs, filterText }) => (
   Object.keys(gifs)
-    .map(key => gifs[key])
+    .map((key) => {
+      const markedGif = {
+        ...gifs[key],
+        isFavorite: !!favoriteGifs[key],
+      };
+      return markedGif;
+    })
     .filter(gif => gif.title.indexOf(filterText) !== -1)
     .map(gif => <Gif object={gif} key={gif.id} />)
 );
